Move ast util function into ast-utils
